# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Nexa REST API
  version: 1.0.0-b01
  summary: Single-node multi-tenant batch AI control plane
  license:
    name: License not declared by B01
    identifier: LicenseRef-Nexa-Unspecified
  description: |
    Contract for browser, CLI, administrator, coordinator-facing worker, and
    local bootstrap operations. Authorization, state transitions, transaction
    semantics, and failure handling are normative in the linked B01 contract
    documents. CLI scopes are exact, non-hierarchical permissions: jobs:read
    covers template/job/session/event/log/result reads; jobs:write covers submit,
    sweep and job controls; artifacts:read and artifacts:write cover tenant
    artifact reads and uploads; tokens:write covers only the caller's token
    metadata lifecycle; admin:read and admin:write cover SYSTEM_ADMIN read and
    mutation operations respectively. Role, tenant membership, ownership, CSRF
    and live authority checks remain mandatory in addition to scope. Unknown
    request fields and duplicate JSON members are rejected.
...
servers:
  - url: https://nexa.example/v1
    description: Deployment-local TLS endpoint; hostname is configuration.
...
paths:
...
  /admin/workers:
    get:
      tags: [Admin]
      operationId: adminListWorkers
      summary: List the configured local worker and health/capacity
      security: [ { browserCookie: [] }, { cliBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/PageSize' } ]
      responses: { '500': { $ref: '#/components/responses/InternalError' }, '503': { $ref: '#/components/responses/DependencyUnavailable' }, '200': { description: Worker page; v1 has one local worker identity., content: { application/json: { schema: { $ref: '#/components/schemas/WorkerPage' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } }
  /admin/workers/{worker_id}:
    get:
      tags: [Admin]
      operationId: adminGetWorker
      summary: Read local worker inventory, incarnation, health, and administrative state
      security: [ { browserCookie: [] }, { cliBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/WorkerId' } ]
      responses: { '400': { $ref: '#/components/responses/BadRequest' }, '500': { $ref: '#/components/responses/InternalError' }, '503': { $ref: '#/components/responses/DependencyUnavailable' }, '200': { description: Worker representation., headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Worker' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
...
  /workers/{worker_id}/incarnations:
    post:
      tags: [Worker]
      operationId: workerCreateIncarnation
      summary: Create and make current a server-generated worker process incarnation
      description: The authenticated configured worker calls this once after taking the local singleton and before heartbeat/reconciliation. Under the Worker row lock the server generates the incarnation ID, increments the monotonic incarnation sequence, marks the prior incarnation ended, makes the new incarnation current in STARTING, and records the idempotent response. Same key and process-start nonce replay the same response; a new committed incarnation makes callbacks from every older incarnation stale.
      security: [ { workerBearer: [] } ]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/WorkerIncarnationCreateRequest' } } } }
      responses:
        '201': { description: New current STARTING incarnation committed., content: { application/json: { schema: { $ref: '#/components/schemas/WorkerIncarnation' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /workers/{worker_id}/reconciliation:
    get:
      tags: [Worker]
      operationId: workerGetReconciliation
      summary: Page through server-known unreleased allocations before worker READY
      description: The cursor is bound to worker identity and current STARTING incarnation. The worker drains all pages and reconciles exact startup/container identities before reporting `reconcile_complete=true`. A LIVE prior-incarnation item is not permission to continue compute; the new incarnation must complete `workerAdoptAttempt`, and only a timely valid adoption acknowledgment may update the existing runner deadline. An unclaimed current-incarnation offer with no journal/container remains eligible for its first poll. A current-incarnation pending claim may replay its original callback after a complete journal and Docker identity scan even while READY is blocked. A succeeded attempt may carry its exact committed completion callback acknowledgment and Result/manifest IDs for terminal reconciliation; worker must match the durable local callback, reservation and committed manifest binding before storing the lost acknowledgment or discarding a pending renewal.
      security: [ { workerBearer: [] } ]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/WorkerIncarnationIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '500': { $ref: '#/components/responses/InternalError' }
        '200':
          description: Bounded reconciliation page ordered by allocation ID.
          content: { application/json: { schema: { $ref: '#/components/schemas/ReconciliationPage' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /workers/{worker_id}/heartbeat:
    post:
      tags: [Worker]
      operationId: workerHeartbeat
      summary: Report worker incarnation, observed health, inventory, and held containers
      security: [ { workerBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/WorkerId' }, { $ref: '#/components/parameters/CallbackId' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/HeartbeatRequest' } } } }
      responses: { '400': { $ref: '#/components/responses/BadRequest' }, '500': { $ref: '#/components/responses/InternalError' }, '200': { description: Heartbeat committed with server DB time and desired admin state., content: { application/json: { schema: { $ref: '#/components/schemas/HeartbeatResponse' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationFailed' }, '503': { $ref: '#/components/responses/DependencyUnavailable' } }
  /workers/{worker_id}/poll:
    post:
      tags: [Worker]
      operationId: workerPollDispatch
      summary: Poll for at most one live dispatch offer
      description: Poll never authorizes execution; claim/start must commit before the runner starts.
      security: [ { workerBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/WorkerId' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PollRequest' } } } }
      responses: { '400': { $ref: '#/components/responses/BadRequest' }, '500': { $ref: '#/components/responses/InternalError' }, '200': { description: 'A committed dispatch offer with an unrevoked lease live at PostgreSQL DB time, or an explicit no-work response.', content: { application/json: { schema: { $ref: '#/components/schemas/PollResponse' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationFailed' }, '503': { $ref: '#/components/responses/DependencyUnavailable' } }
  /attempts/{attempt_id}/claim:
    post:
      tags: [Worker]
      operationId: workerClaimAttempt
      summary: Confirm a dispatch offer for the exact worker incarnation
      security: [ { workerBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/AttemptId' }, { $ref: '#/components/parameters/CallbackId' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AuthorityRequest' } } } }
      responses: { '400': { $ref: '#/components/responses/BadRequest' }, '500': { $ref: '#/components/responses/InternalError' }, '200': { description: Claim acknowledgment and immutable execution context; replay returns the byte-equivalent context snapshot., content: { application/json: { schema: { $ref: '#/components/schemas/ClaimResponse' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationFailed' }, '503': { $ref: '#/components/responses/DependencyUnavailable' } }
  /attempts/{attempt_id}/adopt:
    post:
      tags: [Worker]
      operationId: workerAdoptAttempt
      summary: Adopt an exact live managed container into the current worker incarnation
      description: During STARTING reconciliation only, the current incarnation may atomically transfer an exact live Attempt/Lease from the immediately prior incarnation when the prior Authority is still live, desired state permits continuation, and container ID plus runtime identity digest match the server record and local inspection. Under the same Job/Attempt/Lease/reservation locks, the commit appends the immutable prior-to-current Authority grant lineage and rebinds every still-active checkpoint/result reservation without changing reservation IDs, callback IDs, upload keys, Artifact responses or runner bindings; the response explicitly returns transferred reservations. A stable descriptor-key retry may recover a completed upload response from a lineage predecessor only after current-Authority authentication and exact attempt/allocation/lease/fence plus immutable upload-metadata match. Same adoption callback replays the original complete acknowledgment. Failed, expired, mismatched, local-runner-binding-conflicting, or response-too-late adoption never authorizes execution; the worker stops the container, reports failure/cleanup, and cannot report READY until resolved.
      security: [ { workerBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/AttemptId' }, { $ref: '#/components/parameters/CallbackId' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdoptRequest' } } } }
      responses: { '400': { $ref: '#/components/responses/BadRequest' }, '500': { $ref: '#/components/responses/InternalError' }, '200': { description: Adoption acknowledgment with new Authority and lease values., content: { application/json: { schema: { $ref: '#/components/schemas/AdoptResponse' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationFailed' }, '503': { $ref: '#/components/responses/DependencyUnavailable' } }
  /attempts/{attempt_id}/start:
    post:
      tags: [Worker]
      operationId: workerStartAttempt
      summary: Publish verified container identity and transition to RUNNING
      description: The first committed callback sets DB lease expiry to DB-now plus 45 seconds and stores that acknowledgment. The worker binds the callback's first send-monotonic instant and applies its candidate runner deadline only after a valid acknowledgment arrives before that deadline. Same-callback replay returns the original expiry without extending it; failed or delayed response never starts/rebases compute.
      security: [ { workerBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/AttemptId' }, { $ref: '#/components/parameters/CallbackId' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StartRequest' } } } }
      responses: { '400': { $ref: '#/components/responses/BadRequest' }, '500': { $ref: '#/components/responses/InternalError' }, '200': { description: Start acknowledgment with the initial lease deadline contract., content: { application/json: { schema: { $ref: '#/components/schemas/StartResponse' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationFailed' }, '503': { $ref: '#/components/responses/DependencyUnavailable' } }
  /attempts/{attempt_id}/renew:
    post:
      tags: [Worker]
      operationId: workerRenewAttempt
      summary: Renew a live attempt lease
      description: Before the first send for a callback ID the worker records that callback's immutable local send-monotonic instant. Before the first runner progress frame it sends the lease-only branch `progress_sequence=0, progress=null`, which leaves progress explicitly absent; afterward it sends sequence >=1 and the complete latest snapshot. It may apply `send_monotonic + lease_duration - safety_margin` to the runner only after receiving this valid acknowledgment before that candidate deadline. Failure, uncertainty, delayed response, and same-callback replay never extend or rebase unconfirmed authority.
      security: [ { workerBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/AttemptId' }, { $ref: '#/components/parameters/CallbackId' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RenewRequest' } } } }
      responses: { '400': { $ref: '#/components/responses/BadRequest' }, '500': { $ref: '#/components/responses/InternalError' }, '200': { description: Lease renewed using DB time., content: { application/json: { schema: { $ref: '#/components/schemas/RenewResponse' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationFailed' }, '503': { $ref: '#/components/responses/DependencyUnavailable' } }
...
  /attempts/{attempt_id}/artifacts:
    post:
      tags: [Worker]
      operationId: workerUploadAttemptArtifact
      summary: Stream an attempt-scoped checkpoint, result, or log artifact
      description: The worker supplies full live Authority in headers; cleanup-only authority cannot upload. `application/octet-stream` is the uniform transport and required `X-Artifact-Media-Type` carries validated original media metadata. Idempotency hashes declared media type, size/checksum, artifact kind, normalized attempt path and the full Authority tuple, and current credential/incarnation/live authority are rechecked before replay. The only cross-incarnation replay is a completed upload whose stored Authority is a verified predecessor of current Authority for the same Attempt after successful adoption; attempt/allocation/lease/fence and immutable upload metadata must still match exactly. `X-Artifact-Size` is exact and must equal Content-Length when present or the final chunked byte count. Bytes become a committed Artifact only after the durable blob sequence and metadata commit; interruption/mismatch returns 422 and leaves only expiring staging. The successful or replayed `201 Artifact` response is the sole source of the committed artifact ID/kind/media type/size/checksum binding returned to the trusted runner before it creates a final checkpoint/result manifest.
      security: [ { workerBearer: [] } ]
      parameters:
        - $ref: '#/components/parameters/AttemptId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ArtifactChecksum'
        - $ref: '#/components/parameters/ArtifactSize'
        - $ref: '#/components/parameters/AttemptArtifactKindHeader'
        - $ref: '#/components/parameters/ArtifactMediaTypeHeader'
        - $ref: '#/components/parameters/WorkerIdHeader'
        - $ref: '#/components/parameters/WorkerIncarnationIdHeader'
        - $ref: '#/components/parameters/AllocationIdHeader'
        - $ref: '#/components/parameters/LeaseIdHeader'
        - $ref: '#/components/parameters/JobFenceHeader'
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
        '201':
          description: Durable attempt-scoped artifact committed; publish still requires a separate fenced callback.
          headers: { Location: { schema: { type: string, format: uri-reference } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Artifact' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /attempts/{attempt_id}/result-reservations:
    post:
      tags: [Worker]
      operationId: workerReserveResult
      summary: Reserve the server-generated identity for the unique final result
      description: Exact live Authority reserves one immutable server-generated result ID for its Attempt. Same callback replays the same reservation; a distinct reservation under the same live Authority conflicts. Successful incarnation adoption of that same Attempt atomically rebinds the active reservation to the new Authority and returns it in the adoption response without changing its ID/callback. After Authority is otherwise revoked and no Result exists, only a later authorized Attempt may supersede the abandoned reservation with a new server ID under the Job lock. Abandonment never creates a recognized Result, and completion must present the currently active reservation ID.
      security: [ { workerBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/AttemptId' }, { $ref: '#/components/parameters/CallbackId' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AuthorityRequest' } } } }
      responses: { '201': { description: Final result identity reserved., content: { application/json: { schema: { $ref: '#/components/schemas/ResultReservation' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationFailed' }, '500': { $ref: '#/components/responses/InternalError' }, '503': { $ref: '#/components/responses/DependencyUnavailable' } }
...
components:
  securitySchemes:
    browserCookie: { type: apiKey, in: cookie, name: nexa_session }
    csrfToken: { type: apiKey, in: header, name: X-CSRF-Token }
    cliBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-cli-token
      description: Exact non-hierarchical scopes. `jobs:read` authorizes template/job/session/event/log/result reads; `jobs:write` authorizes submit, sweep and job control; `artifacts:read` authorizes tenant artifact list/download; `artifacts:write` authorizes tenant upload; `tokens:write` authorizes only the caller's token list/create/revoke; `admin:read` and `admin:write` authorize SYSTEM_ADMIN reads and mutations. Current-principal introspection accepts any otherwise-valid token. No scope bypasses membership, ownership, global grant, CSRF, tenant context or object-state checks.
    workerBearer: { type: http, scheme: bearer, bearerFormat: opaque-worker-token }
    bootstrapSecret: { type: apiKey, in: header, name: X-Nexa-Bootstrap-Secret }
...